Moved user/table not found exception to ContextManager and handling only NotFoundException subexception
